feat: Allow decimal input for SpO2 percentage (1 decimal place)
- Change SpO2 percentage from Integer to Numeric(4,1) in database
- Update schema to use float instead of int
- Add step="0.1" to SpO2 form input
- Round to 1 decimal place on submit
- Display with 1 decimal place (e.g., 98.0%)

+++ backend/app/models/spo2_record.py
@@ -5,7 +5,7 @@ SpO2 (Blood Oxygen Saturation) record model.
 from datetime import datetime
 from uuid import UUID, uuid4
 
-from sqlalchemy import Column, DateTime, ForeignKey, Integer, String, Text
+from sqlalchemy import Column, DateTime, ForeignKey, Numeric, String, Text
 from sqlalchemy.dialects.postgresql import UUID as PostgreSQLUUID
 from sqlalchemy.orm import relationship
 
@@ -25,7 +25,7 @@ class SpO2Record(Base):
         index=True,
     )
     recorded_at = Column(DateTime, nullable=False, index=True, comment="When the measurement was taken")
-    percentage = Column(Integer, nullable=False, comment="Blood oxygen saturation (%)")
+    percentage = Column(Numeric(4, 1), nullable=False, comment="Blood oxygen saturation (%)")
     visibility = Column(
         String(20),
         nullable=False,
